requirements.txt
tensorflow
keras 
flask 
opencv-python 
pillow
scikit-learn
numpy
pandas
-e .